refactor(iota-core): move metrics to it's own file
