+++ crates/iota-core/src/authority_server/test_server.rs
@@ -1,3 +1,8 @@
+// Copyright (c) 2021, Facebook, Inc. and its affiliates
+// Copyright (c) Mysten Labs, Inc.
+// Modifications Copyright (c) 2024 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
 use std::{io, sync::Arc};
 
 use anyhow::Result;
